pyproject.toml
[project]
name = "script-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.39.1",
    "networkx>=3.5",
    "pydantic>=2.11.9",
]
